refactored core-tests
