+++ project/Build.scala
@@ -166,7 +166,8 @@ object SummingbirdBuild extends Build {
     summingbirdSpark,
     summingbirdBuilder,
     summingbirdChill,
-    summingbirdExample
+    summingbirdExample,
+    summingbirdCoreTest
   )
 
 
@@ -241,6 +242,7 @@ object SummingbirdBuild extends Build {
     )
   ).dependsOn(
     summingbirdCore % "test->test;compile->compile",
+    summingbirdCoreTest % "test->test",
     summingbirdBatch,
     summingbirdClient
   )
@@ -280,6 +282,7 @@ object SummingbirdBuild extends Build {
     )
   ).dependsOn(
     summingbirdCore % "test->test;compile->compile",
+    summingbirdCoreTest % "test->test",
     summingbirdStorm
   )
 
@@ -321,6 +324,7 @@ object SummingbirdBuild extends Build {
     )
   ).dependsOn(
     summingbirdCore % "test->test;compile->compile",
+    summingbirdCoreTest % "test->test",
     summingbirdChill,
     summingbirdBatchHadoop,
     summingbirdScalding
@@ -403,6 +407,19 @@ object SummingbirdBuild extends Build {
   .settings(sparkAssemblyMergeSettings:_*)
   .dependsOn(
     summingbirdCore % "test->test;compile->compile",
+    summingbirdCoreTest % "test->test",
     summingbirdChill
   )
+
+  lazy val summingbirdCoreTest = module("core-test").settings(
+    parallelExecution in Test := false,
+    libraryDependencies ++=Seq(
+      "junit" % "junit" % "4.11" % "provided",
+      "org.slf4j" % "slf4j-api" % slf4jVersion,
+      "org.scalacheck" %% "scalacheck" % "1.10.0" % "provided",
+      "org.specs2" %% "specs2" % "1.13" % "provided")
+
+  ).dependsOn(
+      summingbirdCore % "test->test;compile->compile"
+    )
 }
