+++ summingbird-core-test/src/test/scala/com/twitter/summingbird/graph/HMapTests.scala
