+++ summingbird-core-test/src/test/scala/com/twitter/summingbird/graph/LiteralTests.scala
